Simplify TestCoordinatorDynamicFiltering tests
